Update 1.0.6
--settings.txt Added
This file keeps default settings for a new line and using character for printing. It can be change with funcs in myFuncs.h

--Updated myFuncs.h
New functions for reading, writing and reseting, working with settings.
1-resetSettings() : This func changes values with defaults.
2-readSettings() : This func reads from settings.txt and returns them as an array form.
3-writeSettings(): This func gets new settings as an input array and writes them into settings.txt

--Updatedstr2stars.h
New global variables added in myFuncs.h. MAX_SPACES for keeping count of space character in a single line. USING_SHAPE for keeping which character will be use to print